+++ pkg/teeth/rawmetadata.go
@@ -56,7 +56,7 @@ type RawMetadataPlatformsItem struct {
 
 const rawMetadataJSONSchema = `
 {
-	"$schema": "http://json-schema.org/draft-07/schema#",
+	"$schema": "https://json-schema.org/draft-07/schema#",
 	"type": "object",
 	"properties": {
 		"format_version": {
@@ -80,12 +80,20 @@ const rawMetadataJSONSchema = `
 				},
 				"author": {
 					"type": "string"
+				},
+				"tags": {
+					"type": "array",
+					"items": {
+						"type": "string",
+						"pattern": "^[a-zA-Z0-9-]+$"
+					}
 				}
 			},
 			"required": [
 				"name",
 				"description",
-				"author"
+				"author",
+				"tags"
 			]
 		},
 		"commands": {
@@ -251,6 +259,7 @@ const rawMetadataJSONSchema = `
 		"info"
 	]
 }
+
 `
 
 func NewRawMetadata(jsonBytes []byte) (RawMetadata, error) {